# tail recursive count crossing