Should reduce memory usage and execution speed. I don't have any benchmarks to demonstrate this, just the fact that it is now tail recursive.