@pkinney 2018-08-31
@whossname Thanks for the PR!

I did some benchmarks based not he current version and your implementation:

Original Version:
```
## PointPolygonBench
benchmark name                                    iterations   average time
Point way outside of Polygon with envelope check      100000   10.38 µs/op
Point way outside of Polygon                          100000   13.65 µs/op
Point outside of Polygon                              100000   16.11 µs/op
Point on vertex of Polygon                            100000   16.21 µs/op
Point inside Polygon                                  100000   17.11 µs/op
Point on edge of Polygon                              100000   19.47 µs/op
MultiPoint / Polygon intersects                        20000   79.61 µs/op
Polygon / MultiPoint contains                          10000   143.33 µs/op
Cities in States with Envelope check                    1000   1214.22 µs/op
Cities in States                                        1000   1718.43 µs/op
Point / Polygon intersects                               500   3262.72 µs/op
```

PR#11:
```
## PointPolygonBench
benchmark name                                    iterations   average time
Point way outside of Polygon with envelope check      100000   10.58 µs/op
Point way outside of Polygon                          100000   14.43 µs/op
Point on vertex of Polygon                            100000   14.43 µs/op
Point outside of Polygon                              100000   16.77 µs/op
Point on edge of Polygon                              100000   17.79 µs/op
Point inside Polygon                                  100000   18.75 µs/op
MultiPoint / Polygon intersects                        20000   108.23 µs/op
Polygon / MultiPoint contains                          10000   164.12 µs/op
Cities in States with Envelope check                    1000   1169.47 µs/op
Cities in States                                        1000   1593.41 µs/op
Point / Polygon intersects                               500   3681.47 µs/op
```

Not a huge difference, but this is just testing execution speed and not memory utilization.

In any case it's a cleaner implementation, so I'll get it merged in and published.  Thanks again!